Upgrade bit vector

+++ include/EliasFano.h
@@ -227,7 +227,7 @@ class EliasFano {
             return ElementPointer(h, positionH, 0, *this);
         }
 
-        [[nodiscard]] uint64_t at(int position) const {
+        [[nodiscard]] uint64_t at(size_t position) const {
             if (rankSelect == nullptr) {
                 throw new std::logic_error("Rank/Select not initialized yet. Missing call to buildRankSelect");
             }
